+++ cleo_sdk/CLEO.h
@@ -462,6 +462,7 @@ DWORD WINAPI CLEO_GetVersion();
 eGameVersion WINAPI CLEO_GetGameVersion();
 
 BOOL WINAPI CLEO_RegisterOpcode(WORD opcode, _pOpcodeHandler callback);
+BOOL WINAPI CLEO_RegisterCommand(const char* commandName, _pOpcodeHandler callback); // uses cleo\.CONFIG\sa.json to obtain opcode number from name
 void WINAPI CLEO_RegisterCallback(eCallbackId id, void* func);